New uploader script with specific version
The previous version of wheel-upload was annoying because it was easy to
specify a pkg-version string which would pick up the version you wanted
and any release candidates - because the release candidates typically
also start with "mypkg-0.12" (e.g. "mypkg-0.12rc1").  To avoid this easy
mistake, do explicit specify of package name and exact version.
Deprecate previous script.